+++ done/sha.c
@@ -42,16 +42,16 @@ void print_sha_inode(struct unix_filesystem *u, struct inode inode, int inr) {
     if (inode.i_mode & IFDIR) {
         puts("no SHA for directories.");
     } else {
-		struct filev6 fv6 = { .u = u, .i_number = inr, .i_node = inode, .offset = 0 };
+		struct filev6 fv6 = { .u = u, .i_number = (uint16_t)inr, .i_node = inode, .offset = 0 };
         unsigned char content[inode_getsectorsize(&inode) + 1];
         content[inode_getsectorsize(&inode)] = '\0';
-		memset(content, 0, inode_getsectorsize(&inode));
+		memset(content, 0, (size_t)inode_getsectorsize(&inode));
         int error = 0;
         while((error = filev6_readblock(&fv6, &content[fv6.offset])) > 0) {
             if (error < 0) {
                 return;
             }
         };
-        print_sha_from_content(content, inode_getsize(&inode));
+        print_sha_from_content(content, (size_t)inode_getsize(&inode));
     }
 }
